feat: add durability (#23)

+++ constants/blocks.go
@@ -2,7 +2,7 @@ package constants
 
 type Block struct {
 	Value int16
-	Meta  byte
+	Meta  uint16
 }
 
 var (
